+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 1.0.0
+- Other configuration for messages. See example
+## 0.0.5
+- More settings in Android part
 ## 0.0.1
 - Simply allows to register notifications channels on Android. That's it.